+++ ReleaseNotes.md
@@ -4,6 +4,7 @@
 **0.00.05 RC**
 - improved detail pane
 - analyze selection
+- split option
 
 **0.00.04**
 - base64 binary and strings